Re #720 Fixed more Gcc 4.6/7 warnings (only Simplix and Simu V3 left)
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@4985 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: a07b5aed70cffbf1eed4d7897c3f37a18b4639f9
Former-commit-id: af736bceb35162f0ef7762d7904435382a787f89

+++ src/modules/userinterface/legacymenu/confscreens/controlconfig.cpp
@@ -645,7 +645,7 @@ ControlMenuInit(void *prevMenu, void *prefHdle, unsigned index, tGearChangeMode 
 	Cmd[i].labelId = GfuiMenuCreateLabelControl(ScrHandle,param,Cmd[i].name);
 	std::string strCmdEdit(Cmd[i].name);
 	strCmdEdit += " button";
-	Cmd[i].Id = GfuiMenuCreateButtonControl(ScrHandle,param,strCmdEdit.c_str(),(void*)i,onPush,NULL,(tfuiCallback)NULL,onFocusLost);
+	Cmd[i].Id = GfuiMenuCreateButtonControl(ScrHandle,param,strCmdEdit.c_str(),(void*)(long)i,onPush,NULL,(tfuiCallback)NULL,onFocusLost);
 	}
     
     /* Steer Sensibility label and associated editbox */
